% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_val.R
\name{square_val}
\alias{square_val}
\title{square a value}
\usage{
square_val(x)
}
\arguments{
\item{x}{a numeric value to be squared}
}
\value{
a number
}
\description{
square a value
}
\examples{
square_val(2)
}
